pyproject.toml
[project]
name = "4blue"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "channels>=4.3.2",
    "channels-redis>=4.3.0",
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
]
